+++ Userland/Applications/PixelPaint/CMakeLists.txt
@@ -9,40 +9,40 @@ compile_gml(PixelPaintWindow.gml PixelPaintWindowGML.h pixel_paint_window_gml)
 compile_gml(EditGuideDialog.gml EditGuideDialogGML.h edit_guide_dialog_gml)
 
 set(SOURCES
-    BrushTool.cpp
-    BucketTool.cpp
-    CloneTool.cpp
     CreateNewImageDialog.cpp
     CreateNewLayerDialog.cpp
     EditGuideDialog.cpp
     EditGuideDialogGML.h
-    EllipseTool.cpp
-    EraseTool.cpp
-    GuideTool.cpp
     Image.cpp
     ImageEditor.cpp
     Layer.cpp
     LayerListWidget.cpp
     LayerPropertiesWidget.cpp
-    LineTool.cpp
     MainWidget.cpp
-    main.cpp
-    MoveTool.cpp
+    Mask.cpp
     PaletteWidget.cpp
-    PenTool.cpp
-    PickerTool.cpp
     PixelPaintWindowGML.h
     ProjectLoader.cpp
-    RectangleTool.cpp
-    RectangleSelectTool.cpp
-    Mask.cpp
     Selection.cpp
-    SprayTool.cpp
-    ToolboxWidget.cpp
     ToolPropertiesWidget.cpp
-    Tool.cpp
-    ZoomTool.cpp
-)
+    ToolboxWidget.cpp
+    Tools/BrushTool.cpp
+    Tools/BucketTool.cpp
+    Tools/CloneTool.cpp
+    Tools/EllipseTool.cpp
+    Tools/EraseTool.cpp
+    Tools/GuideTool.cpp
+    Tools/LineTool.cpp
+    Tools/MoveTool.cpp
+    Tools/PenTool.cpp
+    Tools/PickerTool.cpp
+    Tools/RectangleSelectTool.cpp
+    Tools/RectangleTool.cpp
+    Tools/SprayTool.cpp
+    Tools/Tool.cpp
+    Tools/ZoomTool.cpp
+    main.cpp
+    )
 
 serenity_app(PixelPaint ICON app-pixel-paint)
 target_link_libraries(PixelPaint LibImageDecoderClient LibGUI LibGfx LibFileSystemAccessClient LibConfig)
