+++ Userland/Applications/PixelPaint/ToolboxWidget.cpp
@@ -5,20 +5,20 @@
  */
 
 #include "ToolboxWidget.h"
-#include "BrushTool.h"
-#include "BucketTool.h"
-#include "CloneTool.h"
-#include "EllipseTool.h"
-#include "EraseTool.h"
-#include "GuideTool.h"
-#include "LineTool.h"
-#include "MoveTool.h"
-#include "PenTool.h"
-#include "PickerTool.h"
-#include "RectangleSelectTool.h"
-#include "RectangleTool.h"
-#include "SprayTool.h"
-#include "ZoomTool.h"
+#include "Tools/BrushTool.h"
+#include "Tools/BucketTool.h"
+#include "Tools/CloneTool.h"
+#include "Tools/EllipseTool.h"
+#include "Tools/EraseTool.h"
+#include "Tools/GuideTool.h"
+#include "Tools/LineTool.h"
+#include "Tools/MoveTool.h"
+#include "Tools/PenTool.h"
+#include "Tools/PickerTool.h"
+#include "Tools/RectangleSelectTool.h"
+#include "Tools/RectangleTool.h"
+#include "Tools/SprayTool.h"
+#include "Tools/ZoomTool.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Button.h>
