+++ Userland/Applications/PixelPaint/Tools/BrushTool.cpp
@@ -6,8 +6,8 @@
  */
 
 #include "BrushTool.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Label.h>
