+++ Userland/Applications/PixelPaint/Tools/BrushTool.h
