+++ Userland/Applications/PixelPaint/Tools/BucketTool.h
