+++ Userland/Applications/PixelPaint/Tools/CloneTool.h
