+++ Userland/Applications/PixelPaint/Tools/EllipseTool.cpp
@@ -5,9 +5,9 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include "RectangleTool.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "EllipseTool.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Label.h>
@@ -20,40 +20,37 @@
 
 namespace PixelPaint {
 
-RectangleTool::RectangleTool()
+EllipseTool::EllipseTool()
 {
 }
 
-RectangleTool::~RectangleTool()
+EllipseTool::~EllipseTool()
 {
 }
 
-void RectangleTool::draw_using(GUI::Painter& painter, Gfx::IntPoint const& start_position, Gfx::IntPoint const& end_position, int thickness)
+void EllipseTool::draw_using(GUI::Painter& painter, Gfx::IntPoint const& start_position, Gfx::IntPoint const& end_position, int thickness)
 {
-    Gfx::IntRect rect;
+    Gfx::IntRect ellipse_intersecting_rect;
     if (m_draw_mode == DrawMode::FromCenter) {
         auto delta = end_position - start_position;
-        rect = Gfx::IntRect::from_two_points(start_position - delta, end_position);
+        ellipse_intersecting_rect = Gfx::IntRect::from_two_points(start_position - delta, end_position);
     } else {
-        rect = Gfx::IntRect::from_two_points(start_position, end_position);
+        ellipse_intersecting_rect = Gfx::IntRect::from_two_points(start_position, end_position);
     }
 
     switch (m_fill_mode) {
-    case FillMode::Fill:
-        painter.fill_rect(rect, m_editor->color_for(m_drawing_button));
-        break;
     case FillMode::Outline:
-        painter.draw_rect_with_thickness(rect, m_editor->color_for(m_drawing_button), thickness);
+        painter.draw_ellipse_intersecting(ellipse_intersecting_rect, m_editor->color_for(m_drawing_button), thickness);
         break;
-    case FillMode::Gradient:
-        painter.fill_rect_with_gradient(rect, m_editor->primary_color(), m_editor->secondary_color());
+    case FillMode::Fill:
+        painter.fill_ellipse(ellipse_intersecting_rect, m_editor->color_for(m_drawing_button));
         break;
     default:
         VERIFY_NOT_REACHED();
     }
 }
 
-void RectangleTool::on_mousedown(Layer* layer, MouseEvent& event)
+void EllipseTool::on_mousedown(Layer* layer, MouseEvent& event)
 {
     if (!layer)
         return;
@@ -66,19 +63,19 @@ void RectangleTool::on_mousedown(Layer* layer, MouseEvent& event)
         return;
 
     m_drawing_button = layer_event.button();
-    m_rectangle_start_position = layer_event.position();
-    m_rectangle_end_position = layer_event.position();
+    m_ellipse_start_position = layer_event.position();
+    m_ellipse_end_position = layer_event.position();
     m_editor->update();
 }
 
-void RectangleTool::on_mouseup(Layer* layer, MouseEvent& event)
+void EllipseTool::on_mouseup(Layer* layer, MouseEvent& event)
 {
     if (!layer)
         return;
 
     if (event.layer_event().button() == m_drawing_button) {
         GUI::Painter painter(layer->bitmap());
-        draw_using(painter, m_rectangle_start_position, m_rectangle_end_position, m_thickness);
+        draw_using(painter, m_ellipse_start_position, m_ellipse_end_position, m_thickness);
         m_drawing_button = GUI::MouseButton::None;
         layer->did_modify_bitmap();
         m_editor->update();
@@ -86,39 +83,36 @@ void RectangleTool::on_mouseup(Layer* layer, MouseEvent& event)
     }
 }
 
-void RectangleTool::on_mousemove(Layer* layer, MouseEvent& event)
+void EllipseTool::on_mousemove(Layer*, MouseEvent& event)
 {
-    if (!layer)
-        return;
-
     if (m_drawing_button == GUI::MouseButton::None)
         return;
 
     m_draw_mode = event.layer_event().alt() ? DrawMode::FromCenter : DrawMode::FromCorner;
 
     if (event.layer_event().shift())
-        m_rectangle_end_position = m_rectangle_start_position.end_point_for_aspect_ratio(event.layer_event().position(), 1.0);
+        m_ellipse_end_position = m_ellipse_start_position.end_point_for_aspect_ratio(event.layer_event().position(), 1.0);
     else if (m_aspect_ratio.has_value())
-        m_rectangle_end_position = m_rectangle_start_position.end_point_for_aspect_ratio(event.layer_event().position(), m_aspect_ratio.value());
+        m_ellipse_end_position = m_ellipse_start_position.end_point_for_aspect_ratio(event.layer_event().position(), m_aspect_ratio.value());
     else
-        m_rectangle_end_position = event.layer_event().position();
+        m_ellipse_end_position = event.layer_event().position();
 
     m_editor->update();
 }
 
-void RectangleTool::on_second_paint(Layer const* layer, GUI::PaintEvent& event)
+void EllipseTool::on_second_paint(Layer const* layer, GUI::PaintEvent& event)
 {
     if (!layer || m_drawing_button == GUI::MouseButton::None)
         return;
 
     GUI::Painter painter(*m_editor);
     painter.add_clip_rect(event.rect());
-    auto start_position = editor_stroke_position(m_rectangle_start_position, m_thickness);
-    auto end_position = editor_stroke_position(m_rectangle_end_position, m_thickness);
-    draw_using(painter, start_position, end_position, AK::max(m_thickness * m_editor->scale(), 1));
+    auto preview_start = m_editor->layer_position_to_editor_position(*layer, m_ellipse_start_position).to_type<int>();
+    auto preview_end = m_editor->layer_position_to_editor_position(*layer, m_ellipse_end_position).to_type<int>();
+    draw_using(painter, preview_start, preview_end, AK::max(m_thickness * m_editor->scale(), 1));
 }
 
-void RectangleTool::on_keydown(GUI::KeyEvent& event)
+void EllipseTool::on_keydown(GUI::KeyEvent& event)
 {
     Tool::on_keydown(event);
     if (event.key() == Key_Escape && m_drawing_button != GUI::MouseButton::None) {
@@ -128,7 +122,7 @@ void RectangleTool::on_keydown(GUI::KeyEvent& event)
     }
 }
 
-GUI::Widget* RectangleTool::get_properties_widget()
+GUI::Widget* EllipseTool::get_properties_widget()
 {
     if (!m_properties_widget) {
         m_properties_widget = GUI::Widget::construct();
@@ -152,7 +146,7 @@ GUI::Widget* RectangleTool::get_properties_widget()
         set_primary_slider(&thickness_slider);
 
         auto& mode_container = m_properties_widget->add<GUI::Widget>();
-        mode_container.set_fixed_height(70);
+        mode_container.set_fixed_height(46);
         mode_container.set_layout<GUI::HorizontalBoxLayout>();
         auto& mode_label = mode_container.add<GUI::Label>("Mode:");
         mode_label.set_text_alignment(Gfx::TextAlignment::CenterLeft);
@@ -162,16 +156,12 @@ GUI::Widget* RectangleTool::get_properties_widget()
         mode_radio_container.set_layout<GUI::VerticalBoxLayout>();
         auto& outline_mode_radio = mode_radio_container.add<GUI::RadioButton>("Outline");
         auto& fill_mode_radio = mode_radio_container.add<GUI::RadioButton>("Fill");
-        auto& gradient_mode_radio = mode_radio_container.add<GUI::RadioButton>("Gradient");
 
         outline_mode_radio.on_checked = [&](bool) {
             m_fill_mode = FillMode::Outline;
         };
         fill_mode_radio.on_checked = [&](bool) {
             m_fill_mode = FillMode::Fill;
-        };
-        gradient_mode_radio.on_checked = [&](bool) {
-            m_fill_mode = FillMode::Gradient;
         };
 
         outline_mode_radio.set_checked(true);
