+++ Userland/Applications/PixelPaint/Tools/EllipseTool.h
