+++ Userland/Applications/PixelPaint/Tools/EraseTool.cpp
@@ -6,8 +6,8 @@
  */
 
 #include "EraseTool.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/CheckBox.h>
