+++ Userland/Applications/PixelPaint/Tools/EraseTool.h
