+++ Userland/Applications/PixelPaint/Tools/LineTool.h
