+++ Userland/Applications/PixelPaint/Tools/MoveTool.cpp
@@ -5,13 +5,12 @@
  */
 
 #include "MoveTool.h"
-#include "Image.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "../Image.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/Menu.h>
 #include <LibGUI/Window.h>
-#include <LibGfx/Bitmap.h>
 
 namespace PixelPaint {
