+++ Userland/Applications/PixelPaint/Tools/MoveTool.h
