+++ Userland/Applications/PixelPaint/Tools/PenTool.cpp
@@ -6,8 +6,8 @@
  */
 
 #include "PenTool.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <LibGUI/Action.h>
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/Label.h>
