+++ Userland/Applications/PixelPaint/Tools/PenTool.h
