+++ Userland/Applications/PixelPaint/Tools/PickerTool.cpp
@@ -6,8 +6,8 @@
  */
 
 #include "PickerTool.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <LibGUI/BoxLayout.h>
 #include <LibGUI/CheckBox.h>
