+++ Userland/Applications/PixelPaint/Tools/PickerTool.h
