+++ Userland/Applications/PixelPaint/Tools/RectangleTool.h
