+++ Userland/Applications/PixelPaint/Tools/SprayTool.cpp
@@ -5,8 +5,8 @@
  */
 
 #include "SprayTool.h"
-#include "ImageEditor.h"
-#include "Layer.h"
+#include "../ImageEditor.h"
+#include "../Layer.h"
 #include <AK/Math.h>
 #include <AK/Queue.h>
 #include <LibGUI/Action.h>
