+++ Userland/Applications/PixelPaint/Tools/SprayTool.h
