+++ Userland/Applications/PixelPaint/Tools/Tool.h
