+++ Userland/Applications/PixelPaint/Tools/ZoomTool.h
